requirements.txt
# DataAnalyzer AI - Python Dependencies

# =============================================================================
# Core AI and Language Model
# =============================================================================
google-generativeai>=0.3.0

# =============================================================================
# Data Analysis and Processing
# =============================================================================
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# =============================================================================
# Data Visualization
# =============================================================================
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# =============================================================================
# Web Application Framework
# =============================================================================
streamlit>=1.28.0
pillow>=9.5.0

# =============================================================================
# File Handling and Data Formats
# =============================================================================
openpyxl>=3.1.0
xlsxwriter>=3.1.0
pyarrow>=12.0.0

# =============================================================================
# Environment and Configuration
# =============================================================================
python-dotenv>=1.0.0
click>=8.1.0

# =============================================================================
# Utilities and Helpers
# =============================================================================
pathlib2>=2.3.7
asyncio-compat>=0.1.2

# =============================================================================
# Optional: Development Dependencies
# =============================================================================
# Uncomment for development
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0